requirements.txt
Django==5.2.6
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
python-decouple==3.8
gunicorn==21.2.0